temp save

+++ lib/alternativeproduct.dart
@@ -2,7 +2,7 @@
 //grade image can be local asset, image have to store online
 class AlternativeProduct {
   final String name,image,grade;
-  final int calories;
+  final double calories;
   final double star;
   AlternativeProduct({
     this.name,
